ai-tools/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
transformers==4.36.0
torch==2.1.2
numpy==1.24.3
torchattacks==3.5.1